Add NumAnswers and Addresses to DNS trace events
Parse and return the IP address answers from A and AAAA DNS records.

Since BPF events are fixed-length, set an arbitrary limit on the number
of addresses. Include the total number of answers in the event so
the user can see when addresses were discarded.

For now, we are parsing IP addresses only for DNS answers
that use message compression
(see https://datatracker.ietf.org/doc/html/rfc1035#section-4.1.4).
This makes it much easier to calculate the size of each answer,
since the answer name is always two octets.
I believe most DNS servers use message compression to avoid
repeating the domain name in both the question and answer
sections.

Signed-off-by: Will Daly <will.e.daly@gmail.com>
Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>